requirements.txt
streamlit
numpy
requests
pillow
opencv-python-headless